reorganize js documents
