+++ app/assets/stylesheets/components/_map.scss
@@ -11,5 +11,5 @@
 }
 
 .easyPrintHolder .a3CssClass {
-  //background-image: url(data:image/svg+xml;utf8;base64,PD9...go=);
+  background-image: url(data:image/svg+xml;utf8;base64,PD9...go=);
 }
