+++ app/javascript/actions/ride-color.js
